requirements.txt
pyyaml>=6.0
requests>=2.28.0
python-dotenv>=1.0.0
anthropic>=0.5.0
openai>=1.0.0
google-generativeai>=0.3.0
bert-score>=0.3.12
torch>=1.10.0
transformers>=4.20.0
nltk>=3.6.0
rouge-score>=0.1.2
